Update rehydratoR.R
Added group_start argument.  It takes the split list of IDs and keeps those from group_start and up.  It is useful in case a download is interrupted; with it, the user does not have to redownload tweets already acquired.

Added a line to print estimated time to download all tweets.